feat: Enhance quantization and embedding pipeline with semantic attributes and new quantizers
- Updated exp01_quantization.yaml to include semantic embedding method and feature selector for 'male'.
- Modified exp01.json to reflect changes in embedding method and quantization parameters.
- Introduced SemanticAttributeEmbedder for semantic attribute extraction based on gender.
- Replaced ProductSphericalKMeansQuantizer with GlobalSphericalKMeansQuantizer and added IdentityQuantizer and BioHashingQuantizer.
- Updated seed generation to use int32 for quantized payloads.
- Enhanced IdentitySeedPipeline to support source paths for embeddings.
- Added BioHashing BER test script for evaluating bit error rates on real samples.
- Implemented GenderClassifier for gender detection in embeddings.

+++ code/src/anon_pipeline/components/seed.py
@@ -21,6 +21,6 @@ class HmacSeedGenerator(SeedGenerator):
         self.digest = digest
 
     def generate(self, quantized: np.ndarray) -> str:
-        payload = quantized.astype("float32").tobytes()
+        payload = quantized.astype("int32").tobytes()
         digest = hmac.new(self.secret_key, payload, self.digest)
         return digest.hexdigest()
